backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sonataops-studio-backend"
version = "0.1.0"
description = "SonataOps Studio backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "asyncpg>=0.29.0",
  "clickhouse-connect>=0.8.0",
  "minio>=7.2.7",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "prometheus-client>=0.20.0",
  "opentelemetry-api>=1.26.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-exporter-otlp>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
  "opentelemetry-instrumentation-logging>=0.47b0",
  "llama-index-core>=0.11.10",
  "python-multipart>=0.0.9",
  "orjson>=3.10.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
